update doc.
when a peer starts a leader election, it (1) immediately sets itself a leader by
calling next_round#() function, which sets the lower bits of round_ to this
peer's id; (2) the peer also reads its tail_ to send to remote peers to recover
the log. The (1) and (2) need to happen atomically, because once we start the
leader election, we don't want the tail_ to advance; the only other thread that
is advancing the tail_ is the execute_and_trim thread, which will stop doing so
as soon as a peer becomes a leader.