update doc-v2.org
we should not blindly append instances to our log; if an instance is for a slot
number lower than min_tail that we currently have, which may happen because
between the time we replied to a prepare request and we received an accept, we
may have moved our min_tail forward. it's not a safety or liveness issue, but it
may result in memory leak because we will end up accumulating instances that we
will never trim.