c++: log, log_test, doc: factor out insert from append
Append contains the logic that is used when inserting an instance to the log.
the same logic is also used when in response to a prepare message, a leader
receives many logs from other peers and merges them to form a single log for
replay. extract out the logic to a separate Insert function and implement Append
using the Insert function. The same function will be used for implementing Merge
in the consensus object.