c++: multipaxos: fix race condition on ballot_
NextBallot performs a multistep operation on ballot_ after acquiring the lock,
but the other member functions do not acquire the lock, which can lead to
obtaining a ballot_ value that is inconsistent -- i.e. we may get a ballot_
value after just incrementing the round portion of ballot and before setting the
id portion. Change ballot_ to a non-atomic and use locks for all accessors.