go: multipaxos: add heartbeat in protobuf and its no-grpc implementation
The heartbeat handler follows the grpc handler format in the current version,
except that it doesn't actually use grpc. Other than that, the function logic
is the same and completed. Implementing grpc is not supoosed to affect the code
in the current handler. Additionally, the protobuf is the same as the one in other
two languages except some specific syntax for go. It will be merged into one in
the future.