c++, doc: multipaxos: let heartbeat handler to trim log for the leader too
since we are sending heartbeat to the leader itself as well, we don't need to
run trim_until when the leader receives and computes the new
global_last_executed. instead, we keep track of global_last_executed in a local
variable and keep sending it to all peers, including self, and let the heartbeat
handler trim the log for all peers.