c++: multipaxos: ensure that rpc server is constructed before destructing it
Building and starting the RPC server is a multi-step process, and it is possible
that before rpc_server_ is completely initialized in the Start method, the test
completes and the Stop method called in another thread accesses the same object,
resulting in TSAN violation. Use a condition variable to ensure that Stop method
does not call Shutdown on rpc_server_ before BuildAndStart has returned.