c++: multipaxos: set the ballot from a response only if it is larger than our ballot.
This avoids a ballot going from a higher value to a lower value, which is
possible, for example, if our ballot is 4, and the peers A and B with ballots 5
and 6, respectively, send us responses; first, we receive a response from B and
set our ballot to 6, and then we receive a response from A and set it back to 5.
This problem is avoided with this patch.