c++: multipaxos: factor out send_heartbeat from the heartbeat thread
more specifically, store the state of the heartbeat responses in a
heap-allocated and reference-counted struct (a struct stored in a shared_ptr)
that is shared among threads. as a result, we don't have to care about the order
of thread shutdowns. we first stop the heartbeat thread, which stops any more
heartbeat messages from being generated; since ongoing threadpool threads still
have a reference to the state struct via shared_ptr, the struct does not
disappear when heartbeat thread exits, and we don't get TSAN errors.