c++: multipaxos: reimplement replicate and other functions with the correct logic,
specifically, previously we would atomically check if the peer is leader, using
IsLeader() and then we would use ballot_ with the outgoing requests. This is
wrong because after IsLeader() check we may receive an RPC and become a
follower. Then our outgoing RPCs would use the ballot_ that does not contain us
in its id field. Hence, we need a function that atomically checks if we are a
leader and returns the ballot number. We do this using GetBallotOrLeader()
function, which returns three values; the first one is whether we are a leader,
the second one is whether we are ready assuming we are a leader (we may not be
ready because we're still replaying the log); the last value changes based on
whether we are leader: if we are leader then the last value returned is the
ballot number, otherwise, it is the id of the current leader.