c++: multipaxos: no need for explicitly checking the leadership
we return the merged log if we got responses from the quorum; otherwise, which
means either (1) we are not a leader, or (2) we got all responses but did not
get an ack from quorum, or (3) some of the responses timed out, we return a
nullopt. it is possible that we get a response from quorum and return a merged
log but we stop being a leader right after; in this case the replay of the log
will fail.