c++: multipaxos: no need for explicitly checking the leadership at every step
Once we call NextBallot(), we are the leader, which means we can break out of
the inner loop after that; therefore, with this change we have an unconditional
break at the end of the inner loop. Even if somehow we stop being a leader by
the time we reach break, we will break out of the inner loop, and the first
condition in the outerloop (WaitUntilFollower) will succeed and we will
immediately fall into the inner loop and start listening for heartbeats.

Therefore, since we can stop being a leader at any point in time after calling
NextBallot, doing leadership checks at every step does not make sense. Hence, we
change Replay to return if SendPrepares does not return a merged log (i.e.
stopped being a leader) or if at any point during Replay, SendAccepts fails due
to lost leadership.