c++: multipaxos: function renames
We have two functions for testing if the peer is leader: Isleader(), which
acquires the global lock, and IsLeaderLockless(), which assumes that the lock is
already held by the caller. It appears that internally, we always use
IsLeaderLockless(), and we only use IsLeader() in unit tests. To avoid verbosity
and to explicitly mark functions used just in unit tests, we rename
IsLeaderLockless() to IsLeader() and we rename IsLeader() to IsLeaderTest().

We do the same rename operation for LeaderLockless() and Leader(); i.e., rename
the former to Leader() and the latter to LeaderTest().