all: rename heartbeat -> commit
our heartbeat messages are commit messages in the first place; we just send
commit messages more frequently and use them as heartbeats, so let's name them
with their primary purpose. this also cleanly maps to the three phases of paxos
protocol and the heartbeat stuff does not stand out as an oddity.