c++: client: bugfix: if multipaxos returns ok, then just return
... instead of calling read again because eventually the executor thread will
write the response and call read again. otherwise, we end up calling read twice
and if a client closes the connection first, then both of those reads return
with an error and both try to stop the client; one of these succeeds and the
second one calls client_manager::stop, which tries to remove the non-existent
client from the map.