c++: client: bugfix: avoid calling stop twice on a client
if the server is shut down, then client_manager::stop_all will be called, which
call stop on the clients. however, that will result in an error return from
client::read(), which will try to call stop again on the same client; we add a
check that if the operation was aborted, then it was probably a server-side
shutdown, so stop was already called; therefore, don't call stop again.