c++: log: bugfix: remove invalid assertion
the assertion asserts that a follower should always receive from a leader a
global_last_executed value that is at least as large as its local
global_last_executed value. but this doesn't have to be the case:

1. peer A is leader, receives and executes command
2. peer A sends to other peers gle = 0 le = 1 and receives le = 0
3. other peers respond with le = 0 and then execute slot 1
4. peer A sends to other peers gle = 0 le = 1 and receives le = 1
5. since peer A has received le = 1 from other peers, it sends gle = 1, le = 1
6. peer A itself receives its message and sets gle to 1 but before its message
   reaches remote peers, one of them decides to become a leader due to slow
   message delivery and sends a commit message to peer A with its own local gle,
   which is 0.
7. peer A has its gle = 1 but receives a commit message with gle = 0 and
   assertion fires

just removing the assertion solves the problem since the while loop is not
executed.