c++: multipaxos: avoid accidental leader elections
currently, the random sleep duration is computed as

  commit_interval + random(0, commit_interval)

which can potentially end up being equal to commit_interval; as a result, an
accidental leader election may start even though the leader is still alive.
change the random sleep duration to

  commit_interval + commit_interval / 2 + random(0, commit_interval / 2)

which will guarantee that sleep duration is in the range of 1.5x and 2x
commit_interval.