c++: multipaxos, multipaxos_test: bugfix -- avoid "never ready" leader case
currently, next_ballot sets the ballot_ which makes the peer calling next_ballot
a leader. as a result, in prepare thread, a peer may continue acting as a leader
and start sending out commit messages (heartbeats) even though it hasn't
received a positive response from a quorum. as a result, replay won't get called
and it won't set ready flag to true; and since the peer is leader already, it
will break out of the inner loop in prepare_thread and it won't start a new
leader election, and the other peers won't start a leader election because they
keep receiving commit messages. hence, we get into a situation where leader
never becomes ready.

to solve this, we only set the peer as a leader after it has received an ok
response to a prepare message from the quorum.