c++: log: add set_last_index method
most of the time the last_index_ field of log advances one at a time via calls
to advance_last_index(). however, when a peer falls behind and then attempts and
succeeds in becoming a leader -- thus runs the prepare phase and gets acks from
the quorum -- it receives a large number of instances where the index of the
latest instance is much larger than its current last_index_. The peer now needs
to run the accept phase for all those instances to get the majority approval for
them before committing them, but it also should continue responding to the
*new* incoming requests; to enable the latter, it needs to set its last_index_
method to the highest last_index from the batch to avoid issuing requests with
indexes that are already used, hence this method.

we set last_index_ to the maximum of last_index_ and the argument because it is
possible the peer running the prepare phase may receive quorum responses from
other peers before it receives a response from itself and the other peers may
have lower last_index number.