c++: multipaxos: remove ready field
currently, when a peer becomes leader it does not accept new requests from
clients until it has recovered (replayed) all the instances it has received in
response to prepare message it has sent to peers. this is not a limitation of
multipaxos but our implementation; we were tracking this using ready and setting
the peer to ready state after replay has completed, but it is not necessary;
with this commit, we are setting the index of the peer to the highest index
received in prepare phase and we concurrently replay received instances and
accept new incoming requests.